back/requirements.txt
aiosqlite==0.17.0
alembic==1.10.2
amqp==5.1.1
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
bcrypt==4.0.1
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
cffi==1.15.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==40.0.1
dictdiffer==0.9.0
ecdsa==0.18.0
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
iso8601==1.1.0
Jinja2==3.1.2
kombu==5.2.4
Mako==1.2.4
MarkupSafe==2.1.2
mypy-extensions==1.0.0
numpy==1.24.2
opencv-python==4.7.0.72
packaging==23.0
passlib==1.7.4
pathspec==0.11.1
platformdirs==3.1.1
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.7
pypika-tortoise==0.1.6
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
pytz==2022.7.1
PyYAML==6.0
redis==4.5.3
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.7
SQLAlchemy-Utils==0.40.0
starlette==0.26.1
tomli==2.0.1
tomlkit==0.11.6
typing_extensions==4.5.0
uvicorn==0.21.1
uvloop==0.17.0
vine==5.0.0
watchfiles==0.18.1
wcwidth==0.2.6
websockets==10.4
